Creation and initialisation of stores only during NeoStores construction, remove laziness.
Previously it was possible to construct stores only when they were requested 'in a lazy way' independently from main NeoStores container lifecycle.
That brings a confusion when container are started/stopped on HA setup, since it was possible to construct stores on closed datasource.
To avoid confusion and simplify component lifecycle, but still keep possibility to open only requested stores
now we can enumerate stores that should be opened, in case if we don't need all, and they will be constructed
during container creation.

In case if we will request store on already closed container, or will request not requested store - exception will be thrown.
