+++ community/kernel/src/main/java/org/neo4j/kernel/impl/storemigration/legacystore/v21/propertydeduplication/PropertyDeduplicator.java
@@ -42,8 +42,6 @@ import org.neo4j.kernel.impl.store.record.PropertyRecord;
 import org.neo4j.kernel.impl.store.record.Record;
 import org.neo4j.logging.NullLogProvider;
 
-import static org.neo4j.kernel.impl.store.StoreFactory.SF_LAZY;
-
 public class PropertyDeduplicator
 {
     private final FileSystemAbstraction fileSystem;
@@ -69,7 +67,8 @@ public class PropertyDeduplicator
     {
         final StoreFactory storeFactory =
                 new StoreFactory( fileSystem, workingDir, pageCache, NullLogProvider.getInstance() );
-        try ( NeoStores neoStores = storeFactory.openNeoStores( SF_LAZY ) )
+        try ( NeoStores neoStores = storeFactory.openNeoStores( NeoStores.StoreType.PROPERTY, NeoStores.StoreType
+                .NODE, NeoStores.StoreType.SCHEMA) )
         {
             PropertyStore propertyStore = neoStores.getPropertyStore();
             NodeStore nodeStore = neoStores.getNodeStore();
