+++ community/kernel/src/main/java/org/neo4j/kernel/impl/util/rawstorereader/RsdrMain.java
@@ -84,7 +84,7 @@ public class RsdrMain
         try ( PageCache pageCache = createPageCache( files, config ) )
         {
             StoreFactory factory = openStore( new File( storedir, MetaDataStore.DEFAULT_NAME ), config, pageCache );
-            NeoStores neoStores = factory.openNeoStoresEagerly();
+            NeoStores neoStores = factory.openAllNeoStores();
             interact( neoStores );
         }
     }
