+++ community/kernel/src/main/java/org/neo4j/unsafe/impl/batchimport/store/BatchingNeoStores.java
@@ -63,7 +63,6 @@ import static java.lang.String.valueOf;
 import static org.neo4j.graphdb.factory.GraphDatabaseSettings.dense_node_threshold;
 import static org.neo4j.graphdb.factory.GraphDatabaseSettings.pagecache_memory;
 import static org.neo4j.helpers.collection.MapUtil.stringMap;
-import static org.neo4j.kernel.impl.store.StoreFactory.SF_CREATE;
 
 /**
  * Creator and accessor of {@link NeoStores} with some logic to provide very batch friendly services to the
@@ -182,7 +181,7 @@ public class BatchingNeoStores implements AutoCloseable, NeoStoresSupplier
         {
             StoreFactory storeFactory =
                     new StoreFactory( fileSystem, new File( storeDir ), pageCache, NullLogProvider.getInstance() );
-            try ( NeoStores neoStores = storeFactory.openNeoStores( SF_CREATE ) )
+            try ( NeoStores neoStores = storeFactory.openAllNeoStores( true ) )
             {
                 neoStores.getMetaDataStore();
                 neoStores.getLabelTokenStore();
@@ -200,7 +199,7 @@ public class BatchingNeoStores implements AutoCloseable, NeoStoresSupplier
         BatchingIdGeneratorFactory idGeneratorFactory = new BatchingIdGeneratorFactory( fileSystem );
         StoreFactory storeFactory =
                 new StoreFactory( storeDir, neo4jConfig, idGeneratorFactory, pageCache, fileSystem, logProvider );
-        return storeFactory.openNeoStores( SF_CREATE );
+        return storeFactory.openAllNeoStores( true );
     }
 
     public IoTracer getIoTracer()
