+++ community/kernel/src/test/java/org/neo4j/kernel/impl/api/store/StorePropertyCursorTest.java
@@ -67,7 +67,6 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
-import static org.neo4j.kernel.impl.store.StoreFactory.SF_CREATE;
 
 import static org.neo4j.kernel.impl.locking.LockService.NO_LOCK;
 
@@ -311,7 +310,7 @@ public class StorePropertyCursorTest
             }
             fs.mkdirs( storeDir );
             StoreFactory storeFactory = new StoreFactory( fs, storeDir, pageCache, log );
-            neoStores = storeFactory.openNeoStores( SF_CREATE );
+            neoStores = storeFactory.openAllNeoStores( true );
             propertyStore = neoStores.getPropertyStore();
         }
