+++ community/kernel/src/test/java/org/neo4j/kernel/impl/store/IdGeneratorRebuildFailureEmulationTest.java
@@ -84,7 +84,7 @@ public class IdGeneratorRebuildFailureEmulationTest
         NeoStores neoStores = null;
         try
         {
-            neoStores = factory.openNeoStoresEagerly();
+            neoStores = factory.openAllNeoStores();
             // emulate a failure during rebuild:
             emulateFailureOnRebuildOf( neoStores );
         }
