+++ community/kernel/src/test/java/org/neo4j/kernel/impl/store/RelationshipGroupStoreTest.java
@@ -57,7 +57,6 @@ import static org.hamcrest.Matchers.equalTo;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertThat;
-import static org.neo4j.kernel.impl.store.StoreFactory.SF_CREATE;
 
 public class RelationshipGroupStoreTest
 {
@@ -136,18 +135,18 @@ public class RelationshipGroupStoreTest
     {
         int expectedThreshold = customThreshold != null ? customThreshold : defaultThreshold;
         StoreFactory factory = factory( customThreshold );
-        NeoStores neoStores = factory.openNeoStores( SF_CREATE );
+        NeoStores neoStores = factory.openAllNeoStores( true );
         assertEquals( expectedThreshold, neoStores.getRelationshipGroupStore().getDenseNodeThreshold() );
         neoStores.close();
 
         // Next time we open it it should be the same
-        neoStores = factory.openNeoStoresEagerly();
+        neoStores = factory.openAllNeoStores();
         assertEquals( expectedThreshold, neoStores.getRelationshipGroupStore().getDenseNodeThreshold() );
         neoStores.close();
 
         // Even if we open with a different config setting it should just ignore it
         factory = factory( 999999 );
-        neoStores = factory.openNeoStoresEagerly();
+        neoStores = factory.openAllNeoStores();
         assertEquals( expectedThreshold, neoStores.getRelationshipGroupStore().getDenseNodeThreshold() );
         neoStores.close();
     }
@@ -307,7 +306,7 @@ public class RelationshipGroupStoreTest
         pageCache = pageCacheRule.withInconsistentReads( pageCache, nextReadIsInconsistent );
         StoreFactory factory = factory( null, pageCache );
 
-        try ( NeoStores neoStores = factory.openNeoStores( SF_CREATE ) )
+        try ( NeoStores neoStores = factory.openAllNeoStores( true ) )
         {
             RelationshipGroupStore relationshipGroupStore = neoStores.getRelationshipGroupStore();
             RelationshipGroupRecord record = new RelationshipGroupRecord( 1, 2, 3, 4, 5, 6, true );
