+++ community/kernel/src/test/java/org/neo4j/kernel/impl/store/StoreFactoryTest.java
@@ -42,7 +42,6 @@ import org.neo4j.test.TargetDirectory;
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertThat;
-import static org.neo4j.kernel.impl.store.StoreFactory.SF_CREATE;
 
 public class StoreFactoryTest
 {
@@ -80,7 +79,7 @@ public class StoreFactoryTest
     public void shouldHaveSameCreationTimeAndUpgradeTimeOnStartup() throws Exception
     {
         // When
-        neoStores = storeFactory.openNeoStores( SF_CREATE );
+        neoStores = storeFactory.openAllNeoStores( true );
         MetaDataStore metaDataStore = neoStores.getMetaDataStore();
 
         // Then
@@ -91,7 +90,7 @@ public class StoreFactoryTest
     public void shouldHaveSameCommittedTransactionAndUpgradeTransactionOnStartup() throws Exception
     {
         // When
-        neoStores = storeFactory.openNeoStores( SF_CREATE );
+        neoStores = storeFactory.openAllNeoStores( true );
         MetaDataStore metaDataStore = neoStores.getMetaDataStore();
 
         // Then
@@ -107,7 +106,7 @@ public class StoreFactoryTest
         StoreFactory readOnlyStoreFactory = new StoreFactory( testDirectory.directory( "readOnlyStore" ),
                 new Config( MapUtil.stringMap( GraphDatabaseSettings.read_only.name(), Settings.TRUE ) ),
                 new DefaultIdGeneratorFactory( fs ), pageCache, fs, NullLogProvider.getInstance() );
-        neoStores = readOnlyStoreFactory.openNeoStores( SF_CREATE );
+        neoStores = readOnlyStoreFactory.openAllNeoStores( true );
         long lastClosedTransactionId = neoStores.getMetaDataStore().getLastClosedTransactionId();
 
         // then
@@ -117,7 +116,7 @@ public class StoreFactoryTest
     @Test( expected = StoreNotFoundException.class )
     public void shouldThrowWhenOpeningNonExistingNeoStores()
     {
-        try ( NeoStores neoStores = storeFactory.openNeoStoresEagerly() )
+        try ( NeoStores neoStores = storeFactory.openAllNeoStores() )
         {
             neoStores.getMetaDataStore();
         }
