+++ community/kernel/src/test/java/org/neo4j/unsafe/impl/batchimport/PropertyEncoderStepTest.java
@@ -44,7 +44,7 @@ import static org.mockito.Matchers.anyLong;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.verifyNoMoreInteractions;
-import static org.neo4j.kernel.impl.store.StoreFactory.SF_CREATE;
+
 import static org.neo4j.unsafe.impl.batchimport.Configuration.DEFAULT;
 
 public class PropertyEncoderStepTest
@@ -61,7 +61,7 @@ public class PropertyEncoderStepTest
         pageCache = pageCacheRule.getPageCache( fsRule.get() );
         StoreFactory storeFactory =
                 new StoreFactory( fsRule.get(), storeDir, pageCache, NullLogProvider.getInstance() );
-        neoStores = storeFactory.openNeoStores( SF_CREATE );
+        neoStores = storeFactory.openAllNeoStores( true );
     }
 
     @After
