+++ enterprise/ha/src/main/java/org/neo4j/kernel/ha/transaction/OnDiskLastTxIdGetter.java
@@ -33,15 +33,22 @@ public class OnDiskLastTxIdGetter implements LastTxIdGetter
         this.neoStoresSupplier = neoStoresSupplier;
     }
 
-    // This method is used to construct credentials for election process.
-    // And can be invoked at any moment of instance lifecycle.
-    // It mean that its possible that we will be invoked when neo stores are stopped
-    // (for example while we copy store)
+    /* This method is used to construct credentials for election process.
+     And can be invoked at any moment of instance lifecycle.
+     It mean that its possible that we will be invoked when neo stores are stopped
+     (for example while we copy store) in that case we will return TransactionIdStore.BASE_TX_ID */
     @Override
     public long getLastTxId()
     {
-        TransactionIdStore neoStore = getNeoStores().getMetaDataStore();
-        return neoStore != null ? neoStore.getLastCommittedTransactionId() : TransactionIdStore.BASE_TX_ID;
+        try
+        {
+            TransactionIdStore neoStore = getNeoStores().getMetaDataStore();
+            return neoStore.getLastCommittedTransactionId();
+        }
+        catch ( Throwable e )
+        {
+            return TransactionIdStore.BASE_TX_ID;
+        }
     }
 
     private NeoStores getNeoStores()
